# feat: add `principal_stx_txs` table to speed up `/transfers` endpoint