@blockstack-devops 2022-02-01
:tada: This PR is included in version 2.0.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@stacks/stacks-blockchain-api-types/v/2.0.0)
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@stacks/blockchain-api-client/v/2.0.0)
- [GitHub release](https://github.com/hirosystems/stacks-blockchain-api/releases/tag/v2.0.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: